pyproject.toml
[tool.poetry]
name = "fast-api-utils"
version = "0.1.0"
description = ""
authors = ["ruslan <belyaev@raketa.im>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.95.0"}
pydantic = "^1.10.7"
SQLAlchemy = "1.4.48"
asyncpg = "^0.27.0"
uvicorn = "^0.22.0"
clickhouse-sqlalchemy = "^0.2.4"
factory-boy = "^3.3.0"
clickhouse-connect = "^0.6.10"


[tool.poetry.group.dev.dependencies]
alembic = "^1.10.4"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
